reorganize files for clarity
